Minor types change to match current OpenTX/EdgeTX

+++ radio/src/targets/common/arm/stm32/delays.cpp
@@ -33,7 +33,7 @@
 //     RCC_APB1PeriphClockCmd(RCC_APB1Periph_TIM6, ENABLE);
 // }
 
-void delay_01us(uint16_t nb)
+void delay_01us(uint32_t nb)
 {
     TIM6->SR = 0;
     TIM6->PSC = 0;
@@ -42,7 +42,7 @@ void delay_01us(uint16_t nb)
     TIM6->CR1 |= TIM_CR1_CEN;
     while (!(TIM6->SR & TIM_SR_UIF));
 }
-void delay_us(uint16_t nb)
+void delay_us(uint32_t nb)
 {
     TIM6->SR = 0;
     TIM6->PSC = 0;
@@ -57,14 +57,14 @@ void delaysInit(void)
   DWT->CYCCNT = 0;
   DWT->CTRL |= DWT_CTRL_CYCCNTENA_Msk;
 }
-void delay_01us(uint16_t nb)
+void delay_01us(uint32_t nb)
 {
   volatile uint32_t dwtStart = DWT->CYCCNT;
   volatile uint32_t dwtTotal = (SYSTEM_TICKS_01US * nb) - 10;
   while ((DWT->CYCCNT - dwtStart) < dwtTotal);
 
 }
-void delay_us(uint16_t nb)
+void delay_us(uint32_t nb)
 {
   volatile uint32_t dwtStart = DWT->CYCCNT;
   volatile uint32_t dwtTotal = (SYSTEM_TICKS_1US * nb) - 10;
