+++ radio/src/targets/taranis/board.h
@@ -121,8 +121,8 @@ void init5msTimer();
 extern "C" {
 #endif
 void delaysInit(void);
-void delay_01us(uint16_t nb);
-void delay_us(uint16_t nb);
+void delay_01us(uint32_t nb);
+void delay_us(uint32_t nb);
 void delay_ms(uint32_t ms);
 #ifdef __cplusplus
 }
